Fix overflow in point-set distance norms
